refactor(HubPoolClient): Reverse ConfigStoreClient dependencies
The HubPoolClient will need to read from the ConfigStoreClient, and so
far there don't seem to be any hard reasons for why the
ConfigStoreClient needs a HubPoolClient instance. Therefore, migrate the
necessary parts of the ConfigStoreClient to the HubPoolClient, and
supply the HubPoolClient with a ConfigStoreClient during instantiation.
